Backend/requirements.txt
fastapi
uvicorn
pymysql==1.1.2
pydantic==2.4.2
sqlalchemy
cloudinary==1.44.1
PyMuPDF== 1.26.5
python-multipart==0.0.20
fastapi-mail==1.6.1
slowapi==0.1.9
redis==7.1.0
matplotlib==3.8.2